# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cattle API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/animal": {
            "get": {
                "summary": "List all animals",
                "operationId": "23df7e20e1400d97781d01622f3e0046",
                "responses": {
                    "200": {
                        "description": "A list of animals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "code": {
                                                "type": "string"
                                            },
                                            "milk": {
                                                "type": "number"
                                            },
                                            "food": {
                                                "type": "number"
                                            },
                                            "weight": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "born": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "shooted_down": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token has expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token has expired"
                                        },
                                        "exception": {
                                            "type": "string",
                                            "example": "Symfony\\\\Component\\\\HttpKernel\\\\Exception\\\\UnauthorizedHttpException"
                                        },
                                        "file": {
                                            "type": "string",
                                            "example": "/path/to/file.php"
                                        },
                                        "line": {
                                            "type": "integer",
                                            "example": 74
                                        },
                                        "trace": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "file": "/path/to/file.php",
                                                    "line": 30,
                                                    "function": "authenticate"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/animal/search": {
            "get": {
                "summary": "Get animal by id or by code",
                "operationId": "5bb63f631939c98480f725a76c31c20f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Code of the animal to search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of the animal to search",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animal details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Unique identifier for the animal",
                                            "type": "integer",
                                            "format": "int"
                                        },
                                        "code": {
                                            "description": "Animal's code",
                                            "type": "string"
                                        },
                                        "milk": {
                                            "description": "Daily milk production in liters",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "food": {
                                            "description": "Daily food consumption in kilograms",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "weight": {
                                            "description": "Weight of the animal in arrobas",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "born": {
                                            "description": "Date of birth of the animal",
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "shooted_down": {
                                            "description": "Indicates if the animal has been shot down (1 for yes, 0 for no)",
                                            "type": "integer",
                                            "format": "int"
                                        },
                                        "created_at": {
                                            "description": "Date when the animal was created",
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "updated_at": {
                                            "description": "Date when the animal was last updated",
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token has expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token has expired"
                                        },
                                        "exception": {
                                            "type": "string",
                                            "example": "Symfony\\\\Component\\\\HttpKernel\\\\Exception\\\\UnauthorizedHttpException"
                                        },
                                        "file": {
                                            "type": "string",
                                            "example": "/path/to/file.php"
                                        },
                                        "line": {
                                            "type": "integer",
                                            "example": 74
                                        },
                                        "trace": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "file": "/path/to/file.php",
                                                    "line": 30,
                                                    "function": "authenticate"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}